style(styles-directory): creating styles directory
