+++ jest.config.js
@@ -12,7 +12,8 @@ const customJestConfig = {
     'src/**/*.js(x)?',
     '!src/**/*stories.js(x)?',
     '!src/**/mock.js',
-    '!src/app/**/*.js(x)?'
+    '!src/app/**/*.js(x)?',
+    '!src/styles/**/*.js(x)?'
   ],
   snapshotSerializers: [
     '@emotion/jest/serializer' /* if needed other snapshotSerializers should go here */
