--- src/components/ThemeRegistry/theme.js
@@ -1,30 +1,0 @@
-import { Montserrat } from 'next/font/google'
-import { createTheme } from '@mui/material/styles'
-
-const montserrat = Montserrat({
-  weight: ['400', '500', '700'],
-  subsets: ['latin'],
-  display: 'swap'
-})
-
-const theme = createTheme({
-  palette: {
-    mode: 'light'
-  },
-  typography: {
-    fontFamily: montserrat.style.fontFamily
-  },
-  components: {
-    MuiAlert: {
-      styleOverrides: {
-        root: ({ ownerState }) => ({
-          ...(ownerState.severity === 'info' && {
-            backgroundColor: '#006aff'
-          })
-        })
-      }
-    }
-  }
-})
-
-export default theme
