+++ src/styles/ThemeRegistry.js
@@ -3,6 +3,7 @@ import * as React from 'react'
 import { ThemeProvider } from '@mui/material/styles'
 import CssBaseline from '@mui/material/CssBaseline'
 import NextAppDirEmotionCacheProvider from './EmotionCache'
+import Global from './global'
 import theme from './theme'
 
 export default function ThemeRegistry({ children }) {
@@ -11,6 +12,7 @@ export default function ThemeRegistry({ children }) {
       <ThemeProvider theme={theme}>
         {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
         <CssBaseline />
+        <Global />
         {children}
       </ThemeProvider>
     </NextAppDirEmotionCacheProvider>
